+++ src/core/injections/blind/techniques/time_based/tb_handler.py
@@ -18,7 +18,6 @@ import sys
 import time
 import string
 import random
-import base64
 import urllib
 import urllib2
 
@@ -144,10 +143,9 @@ def tb_injection_handler(url, delay, filename, http_request_method, url_time_res
 
               # Whitespace fixation
               payload = re.sub(" ", whitespace, payload)
-
-              if settings.TAMPER_SCRIPTS['base64encode']:
-                from src.core.tamper import base64encode
-                payload = base64encode.encode(payload)
+              
+              # Check for base64 / hex encoding
+              payload = checks.perform_payload_encoding(payload)
 
               # Check if defined "--verbose" option.
               if settings.VERBOSITY_LEVEL == 1: