+++ src/core/injections/blind/techniques/time_based/tb_injector.py
@@ -197,8 +197,8 @@ def injection(separator, maxlen, TAG, cmd, prefix, suffix, whitespace, delay, ht
     # Whitespace fixation
     payload = re.sub(" ", whitespace, payload)
 
-    if settings.TAMPER_SCRIPTS['base64encode']:
-      payload = base64.b64encode(payload)
+    # Check for base64 / hex encoding
+    payload = checks.perform_payload_encoding(payload)
 
     # Check if defined "--verbose" option.
     if settings.VERBOSITY_LEVEL == 1:
@@ -282,9 +282,9 @@ def injection(separator, maxlen, TAG, cmd, prefix, suffix, whitespace, delay, ht
 
         # Whitespace fixation
         payload = re.sub(" ", whitespace, payload)
-
-        if settings.TAMPER_SCRIPTS['base64encode']:
-          payload = base64.b64encode(payload)
+        
+        # Check for base64 / hex encoding
+        payload = checks.perform_payload_encoding(payload)
 
         # Check if defined "--verbose" option.
         if settings.VERBOSITY_LEVEL == 1:
@@ -383,8 +383,8 @@ def false_positive_check(separator, TAG, cmd, whitespace, prefix, suffix, delay,
     # Whitespace fixation
     payload = re.sub(" ", whitespace, payload)
 
-    if settings.TAMPER_SCRIPTS['base64encode']:
-      payload = base64.b64encode(payload)
+    # Check for base64 / hex encoding
+    payload = checks.perform_payload_encoding(payload)
 
     # Check if defined "--verbose" option.
     if settings.VERBOSITY_LEVEL == 1:
@@ -447,8 +447,8 @@ def false_positive_check(separator, TAG, cmd, whitespace, prefix, suffix, delay,
         # Whitespace fixation
         payload = re.sub(" ", whitespace, payload)
 
-        if settings.TAMPER_SCRIPTS['base64encode']:
-          payload = base64.b64encode(payload)
+        # Check for base64 / hex encoding
+        payload = checks.perform_payload_encoding(payload)
 
         # Check if defined "--verbose" option.
         if settings.VERBOSITY_LEVEL == 1: