+++ src/core/injections/results_based/techniques/classic/cb_handler.py
@@ -19,7 +19,6 @@ import sys
 import time
 import string
 import random
-import base64
 import urllib
 import urllib2
 import HTMLParser
@@ -129,10 +128,9 @@ def cb_injection_handler(url, delay, filename, http_request_method):
 
               # Whitespace fixation
               payload = re.sub(" ", whitespace, payload)
-
-              if settings.TAMPER_SCRIPTS['base64encode']:
-                from src.core.tamper import base64encode
-                payload = base64encode.encode(payload)
+              
+              # Check for base64 / hex encoding
+              payload = checks.perform_payload_encoding(payload)
 
               # Check if defined "--verbose" option.
               if settings.VERBOSITY_LEVEL == 1: