+++ src/core/injections/results_based/techniques/classic/cb_injector.py
@@ -20,7 +20,6 @@ import time
 import json
 import string
 import random
-import base64
 import urllib
 import urllib2
 import HTMLParser
@@ -167,10 +166,9 @@ def injection(separator, TAG, cmd, prefix, suffix, whitespace, http_request_meth
 
     # Whitespace fixation
     payload = re.sub(" ", whitespace, payload)
-
-    if settings.TAMPER_SCRIPTS['base64encode']:
-      from src.core.tamper import base64encode
-      payload = base64encode.encode(payload)
+    
+    # Check for base64 / hex encoding
+    payload = checks.perform_payload_encoding(payload)
 
     # Check if defined "--verbose" option.
     if settings.VERBOSITY_LEVEL >= 1: