+++ src/core/injections/semiblind/techniques/file_based/fb_injector.py
@@ -151,9 +151,15 @@ def injection(separator, payload, TAG, cmd, prefix, suffix, whitespace, http_req
     # Whitespace fixation
     payload = re.sub(" ", whitespace, payload)
 
+    # Encode payload to base64 format.
     if settings.TAMPER_SCRIPTS['base64encode']:
       from src.core.tamper import base64encode
-      payload = base64encode.encode(payload)  
+      payload = base64encode.encode(payload)
+
+    # Encode payload to hex format.
+    elif settings.TAMPER_SCRIPTS['hexencode']:
+      from src.core.tamper import hexencode
+      payload = hexencode.encode(payload)  
 
     # Check if defined "--verbose" option.
     if settings.VERBOSITY_LEVEL >= 1: