+++ src/core/injections/semiblind/techniques/file_based/fb_payloads.py
@@ -60,7 +60,8 @@ def decision_alter_shell(separator, TAG, OUTPUT_TEXTFILE):
      settings.CUSTOM_HEADER_INJECTION == True :
     payload = payload.replace("\n", separator)
   else:
-    if not settings.TAMPER_SCRIPTS['base64encode']:
+    if not settings.TAMPER_SCRIPTS['base64encode'] and \
+       not settings.TAMPER_SCRIPTS['hexencode']:
       if settings.TARGET_OS != "win":
         payload = payload.replace("\n","%0d")
 
@@ -119,7 +120,8 @@ def cmd_execution_alter_shell(separator, cmd, OUTPUT_TEXTFILE):
      settings.CUSTOM_HEADER_INJECTION == True:
     payload = payload.replace("\n", separator)
   else:
-    if not settings.TAMPER_SCRIPTS['base64encode']:
+    if not settings.TAMPER_SCRIPTS['base64encode'] and \
+       not settings.TAMPER_SCRIPTS['hexencode']:
       if settings.TARGET_OS != "win":
         payload = payload.replace("\n","%0d")